ONL-4423:refactor: move files from sass directory to styles and apply tailwind classes
