+++ src/scss/components/_index.scss
@@ -1,6 +1,5 @@
 @import '../settings/index';
 @import '../tools/index';
-@import 'ec-action-list/ec-action-list';
 @import 'ec-btn/ec-btn';
 @import 'ec-grid/ec-grid';
 @import 'v-tooltip';
