--- src/scss/components/ec-action-list/_ec-action-list-item.scss
@@ -1,13 +1,0 @@
-.ec-action-list-item {
-  @include shape-border-radius;
-  @include border-level-6;
-
-  padding: 16px;
-  margin: 4px;
-  cursor: pointer;
-
-  &:hover,
-  &:active {
-    background-color: $level-7-backgrounds;
-  }
-}
