+++ src/styles/components/ec-action-list/ec-action-list-group.css
@@ -1,7 +1,7 @@
 .ec-action-list-group {
-  display: flex;
+  @apply tw-flex;
 
   &--column {
-    flex-direction: column;
+    @apply tw-flex-col;
   }
 }
