+++ src/styles/components/index.css
@@ -1,2 +1,3 @@
 @import 'tailwindcss/components';
 @import 'ec-card/ec-card';
+@import 'ec-action-list/ec-action-list';
